requirements.txt
functions-framework==3.*
requests==2.31.0
google-cloud-secret-manager==2.16.0
python-dotenv==1.0.1
google-api-python-client==2.121.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
